requirements.txt
pyserial
influxdb-client
requests